@@ -8,35 +8,35 @@
         "lint": "ng lint"
     },
     "dependencies": {
-        "@angular/animations": "~17.3.1",
-        "@angular/common": "~17.3.1",
-        "@angular/compiler": "~17.3.1",
-        "@angular/core": "~17.3.1",
-        "@angular/forms": "~17.3.1",
-        "@angular/platform-browser": "~17.3.1",
-        "@angular/platform-browser-dynamic": "~17.3.1",
-        "@angular/router": "~17.3.1",
-        "@ngx-translate/core": "~15.0.0",
-        "@ngx-translate/http-loader": "~8.0.0",
+        "@angular/animations": "~18.2.9",
+        "@angular/common": "~18.2.9",
+        "@angular/compiler": "~18.2.9",
+        "@angular/core": "~18.2.9",
+        "@angular/forms": "~18.2.9",
+        "@angular/platform-browser": "~18.2.9",
+        "@angular/platform-browser-dynamic": "~18.2.9",
+        "@angular/router": "~18.2.9",
+        "@ngx-translate/core": "~16.0.4",
+        "@ngx-translate/http-loader": "~16.0.1",
         "rxjs": "~7.8.1",
         "tslib": "~2.6.2",
         "zone.js": "~0.14.4"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "~17.3.3",
-        "@angular-eslint/builder": "~17.3.0",
-        "@angular-eslint/eslint-plugin": "~17.3.0",
-        "@angular-eslint/eslint-plugin-template": "~17.3.0",
-        "@angular-eslint/schematics": "~17.3.0",
-        "@angular-eslint/template-parser": "~17.3.0",
-        "@angular/cli": "~17.3.1",
-        "@angular/compiler-cli": "~17.3.1",
+        "@angular-devkit/build-angular": "~18.2.19",
+        "@angular-eslint/builder": "~18.4.3",
+        "@angular-eslint/eslint-plugin": "~18.4.3",
+        "@angular-eslint/eslint-plugin-template": "~18.4.3",
+        "@angular-eslint/schematics": "~18.4.3",
+        "@angular-eslint/template-parser": "~18.4.3",
+        "@angular/cli": "~18.2.19",
+        "@angular/compiler-cli": "~18.2.9",
         "@types/node": "~20.11.30",
         "@typescript-eslint/eslint-plugin": "7.2.0",
         "@typescript-eslint/parser": "7.2.0",
         "eslint": "~8.57.0",
         "eslint-plugin-deprecation": "~2.0.0",
         "ts-node": "~10.9.1",
-        "typescript": "~5.4.3"
+        "typescript": "~5.5.4"
     }
 }
